pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openbooru"
readme = "README.md"
requires-python = ">=3.10"
license = "GPL-2.0"
authors = [
    { name = "Ben Brady", email = "benbradybusiness@gmail.com" },
]
dependencies = [
    "wheel",
    "requests",
    "requests-futures",
    "numpy",
    "tqdm",
    "anyio",
    "coloredlogs",
    "iso639",
    "python-multipart",
    "cachetools",
    "pymongo",
    "jinja2",
    "boto3",
    "pickledb",

    # fastpi
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "fastapi-cache2[memcache]",

    # Encoding
    "pillow",
    "ffmpeg-python",
    "python-magic",
    "imagehash",

    # Cryptography
    "cryptography",
    "pyotp",
    "pyjwt",
    "passlib[argon2]>=1.7.4",
    "zxcvbn",
    "argon2-cffi",
    "password-strength",

    # Parsing
    "python-box",
    "ujson",
    "ijson",
    "bs4",
    "lxml",
    "hydrus_api",
    "pytumblr",
    "tweepy>=4.10.0",
    "e621==0.0.6",
    # 0.0.7 causes errors
    "pytube",
]
classifiers = [
    "Operating System :: OS Independent",
    "Topic :: Internet",
    "Topic :: Software Development",
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Internet :: WWW/HTTP",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/TheOpenBooru/api"

[tool.hatch.version]
path = "openbooru/__init__.py"


[project.scripts]
openbooru = "openbooru.app:run"


[options.extras_require]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-timeout",
    "httpx",
    "schemathesis",
]

[tool.black]
line-length = 80
target-version = ["py310"]
include = '\.pyi?$'
exclude-files = "__init__.py"

[tool.mypy]
ignore_missing_imports = true